+++ src/custom/hooks/useSwapCallback.ts
@@ -8,9 +8,9 @@ import useENS from '@src/hooks/useENS'
 import { useMemo } from 'react'
 import useTransactionDeadline from '@src/hooks/useTransactionDeadline'
 import { BigNumber } from 'ethers'
-import { useAddPendingOrder } from '../state/operator/hooks'
-import { AddPendingOrderParams, OrderCreation, OrderID, OrderKind, OrderStatus } from '../state/operator/actions'
 import { isAddress, shortenAddress } from '@src/utils'
+import { AddPendingOrderParams, OrderCreation, OrderID, OrderKind, OrderStatus } from '../state/orders/actions'
+import { useAddPendingOrder } from '../state/orders/hooks'
 
 interface PostOrderParams {
   account: string
@@ -108,9 +108,9 @@ async function postOrder(params: PostOrderParams): Promise<string> {
       ...unsignedOrder,
       creationTime,
       signature,
-      status: OrderStatus.PENDING
+      status: OrderStatus.PENDING,
+      summary
     }
-    //summary
   })
 
   console.log('[useSwapCallback] TODO: Add summary also to new pendingOrder', summary)
