+++ src/custom/state/operator/reducer.ts
@@ -1,5 +1,5 @@
 import { createReducer } from '@reduxjs/toolkit'
-import { updateTip, clearTip, Tip, Order, addPendingOrder } from './actions'
+import { updateTip, clearTip, Tip } from './actions'
 
 export interface TipObject {
   token: string // token address
@@ -10,22 +10,15 @@ export interface TipObject {
 export type TipsMap = Record<string, TipObject>
 
 export interface OperatorState {
-  readonly pendingOrders: Order[]
-  readonly pastOrders: Order[]
   readonly tipsMap: Partial<TipsMap>
 }
 
 const initialState: OperatorState = {
-  pendingOrders: [],
-  pastOrders: [],
   tipsMap: {}
 }
 
 export default createReducer(initialState, builder =>
   builder
-    .addCase(addPendingOrder, (state, action) => {
-      console.log('TODO: Add pending orders', action)
-    })
     .addCase(updateTip, (state, action) => {
       const { token, tip } = action.payload
       state.tipsMap[token] = { tip, token }
