+++ src/custom/state/orders/actions.ts
@@ -1,21 +1,22 @@
 import { createAction } from '@reduxjs/toolkit'
 import { ChainId } from '@uniswap/sdk'
 
-enum OrderKind {
+export enum OrderKind {
   SELL = 'sell',
   BUY = 'buy'
 }
 
 // posted to /api/v1/orders on Order creation
 // serializable, so no BigNumbers
+//  See https://protocol-rinkeby.dev.gnosisdev.com/api/
 export interface OrderCreation {
   sellToken: string // address, without '0x' prefix
   buyToken: string // address, without '0x' prefix
   sellAmount: string // in atoms
   buyAmount: string // in atoms
-  validTo: number // unix timestamp, seconds, use new Date(validTo * 1000)
+  validTo: number // uint32. unix timestamp, seconds, use new Date(validTo * 1000)
   appData: number // arbitrary identifier sent along with the order
-  tip: string // in atoms
+  feeAmount: string // in atoms
   orderType: OrderKind
   partiallyFillable: boolean
   signature: string // 65 bytes encoded as hex without `0x` prefix. v + r + s from the spec
@@ -48,7 +49,13 @@ export interface OrderFromApi extends OrderCreation {
  */
 export type OrderID = string
 
-export const addPendingOrder = createAction<{ id: OrderID; chainId: ChainId; order: Order }>('order/addPendingOrder')
+export interface AddPendingOrderParams {
+  id: OrderID
+  chainId: ChainId
+  order: Order
+}
+
+export const addPendingOrder = createAction<AddPendingOrderParams>('order/addPendingOrder')
 export const removeOrder = createAction<{ id: OrderID; chainId: ChainId }>('order/removeOrder')
 //                                                                        fulfillmentTime from event timestamp
 export const fulfillOrder = createAction<{ id: OrderID; chainId: ChainId; fulfillmentTime: string }>(
